chromium: Drop the root-profile PACKAGECONFIG setting.
It's been broken since https://codereview.chromium.org/2578483002, which
made the code fail unless --no-sandbox is specified on the command line.

Instead of fighting an uphill battle by keeping the patch around, just tell
users to run the chromium binary with --no-sandbox when testing, or launch
it with a separate, unprivileged user.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>